refactor(idp-basic): Inline JDBC session init and fix naming
Merge IdpJdbcSessionSupport into IdpJDBCBackend and rename userNames
to usernames across IdP group model APIs.
