+++ plugins/idp-basic/src/main/java/org/apache/gravitino/idp/model/IdpGroup.java
@@ -35,5 +35,5 @@ public interface IdpGroup {
    *
    * @return The usernames of the built-in IdP group.
    */
-  List<String> userNames();
+  List<String> usernames();
 }
